feat(agent): 补齐 AgentConfig 缺失配置字段
- 以 src/agent/config.rs 为唯一定义位置，补齐设计文档要求的全部配置块
- 新增 SubagentsConfig 子结构（allow_agents/require_agent_id/max_spawn_depth/max_children 等）
- BootstrapMode 添加 Serialize/Deserialize，支持 JSON 序列化
- 删除 src/config/agents/types.rs 中的重复 AgentConfig，AgentsConfig 改为引用 agent::config::AgentConfig
- 适配 validation.rs/provider.rs/directory.rs 等所有调用方
- 新增单元测试覆盖新字段默认值、序列化 round-trip、JSON 反序列化

Source: issue #786
Type: feat
