+++ src/config/agents/types.rs
@@ -1,33 +1,9 @@
 //! Agent configuration types
 //!
-//! Provides AgentConfig and AgentsConfig structs.
+//! Provides AgentsConfig struct.
+//! AgentConfig is defined in `crate::agent::config`.
 
 use serde::{Deserialize, Serialize};
-
-/// Agent definition from JSON config
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct AgentConfig {
-    /// Agent identifier
-    pub name: String,
-    /// Model to use for this agent
-    pub model: String,
-    /// Optional human-readable persona description
-    #[serde(default)]
-    pub persona: String,
-    /// Maximum iterations for this agent
-    #[serde(default = "default_max_iterations")]
-    pub max_iterations: u32,
-    /// Timeout in minutes
-    #[serde(default)]
-    pub timeout_minutes: Option<u32>,
-    /// Optional parent agent for delegation
-    #[serde(default)]
-    pub parent: Option<String>,
-}
-
-fn default_max_iterations() -> u32 {
-    100
-}
 
 /// Wrapper for the entire agents.json file
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -35,7 +11,7 @@ pub struct AgentsConfig {
     /// Config version
     pub version: String,
     /// List of agent definitions
-    pub agents: Vec<AgentConfig>,
+    pub agents: Vec<crate::agent::config::AgentConfig>,
 }
 
 impl Default for AgentsConfig {
@@ -50,55 +26,56 @@ impl Default for AgentsConfig {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::agent::config::{AgentConfig, AgentConfigState};
+
+    fn make_agent(name: &str, model: &str) -> AgentConfig {
+        AgentConfig {
+            id: name.to_string(),
+            name: name.to_string(),
+            model: Some(model.to_string()),
+            ..Default::default()
+        }
+    }
 
     #[test]
     fn test_agent_config_deserialize_full() {
         let json = r#"{
+            "id": "test-id",
             "name": "test-agent",
             "model": "gpt-4",
-            "persona": "A test agent",
-            "max_iterations": 50,
-            "timeout_minutes": 30,
-            "parent": "parent-agent"
+            "created_at": "2026-01-01T00:00:00Z"
         }"#;
         let config: AgentConfig = serde_json::from_str(json).unwrap();
+        assert_eq!(config.id, "test-id");
         assert_eq!(config.name, "test-agent");
-        assert_eq!(config.model, "gpt-4");
-        assert_eq!(config.persona, "A test agent");
-        assert_eq!(config.max_iterations, 50);
-        assert_eq!(config.timeout_minutes, Some(30));
-        assert_eq!(config.parent, Some("parent-agent".to_string()));
+        assert_eq!(config.model.as_deref(), Some("gpt-4"));
     }
 
     #[test]
     fn test_agent_config_deserialize_minimal() {
         let json = r#"{
+            "id": "min-id",
             "name": "minimal",
-            "model": "claude-3"
+            "created_at": "2026-01-01T00:00:00Z"
         }"#;
         let config: AgentConfig = serde_json::from_str(json).unwrap();
+        assert_eq!(config.id, "min-id");
         assert_eq!(config.name, "minimal");
-        assert_eq!(config.model, "claude-3");
-        assert_eq!(config.persona, "");
-        assert_eq!(config.max_iterations, 100);
-        assert_eq!(config.timeout_minutes, None);
-        assert_eq!(config.parent, None);
+        assert_eq!(config.model, None);
     }
 
     #[test]
     fn test_agent_config_serialize_roundtrip() {
         let config = AgentConfig {
+            id: "rt".to_string(),
             name: "rt".to_string(),
-            model: "m".to_string(),
-            persona: "p".to_string(),
-            max_iterations: 10,
-            timeout_minutes: Some(5),
-            parent: None,
+            model: Some("m".to_string()),
+            ..Default::default()
         };
         let json = serde_json::to_string(&config).unwrap();
         let parsed: AgentConfig = serde_json::from_str(&json).unwrap();
+        assert_eq!(parsed.id, config.id);
         assert_eq!(parsed.name, config.name);
-        assert_eq!(parsed.max_iterations, config.max_iterations);
     }
 
     #[test]
@@ -113,33 +90,26 @@ mod tests {
         let json = r#"{
             "version": "2.0",
             "agents": [
-                {"name": "a1", "model": "m1"},
-                {"name": "a2", "model": "m2", "max_iterations": 200}
+                {"id": "a1", "name": "a1", "model": "m1", "created_at": "2026-01-01T00:00:00Z"},
+                {"id": "a2", "name": "a2", "model": "m2", "created_at": "2026-01-01T00:00:00Z"}
             ]
         }"#;
         let config: AgentsConfig = serde_json::from_str(json).unwrap();
         assert_eq!(config.version, "2.0");
         assert_eq!(config.agents.len(), 2);
-        assert_eq!(config.agents[0].max_iterations, 100);
-        assert_eq!(config.agents[1].max_iterations, 200);
+        assert_eq!(config.agents[0].model.as_deref(), Some("m1"));
+        assert_eq!(config.agents[1].model.as_deref(), Some("m2"));
     }
 
     #[test]
     fn test_agents_config_serialize_roundtrip() {
         let config = AgentsConfig {
             version: "1.0".to_string(),
-            agents: vec![AgentConfig {
-                name: "x".to_string(),
-                model: "y".to_string(),
-                persona: String::new(),
-                max_iterations: 100,
-                timeout_minutes: None,
-                parent: Some("z".to_string()),
-            }],
+            agents: vec![make_agent("x", "y")],
         };
         let json = serde_json::to_string(&config).unwrap();
         let parsed: AgentsConfig = serde_json::from_str(&json).unwrap();
         assert_eq!(parsed.agents.len(), 1);
-        assert_eq!(parsed.agents[0].parent.as_deref(), Some("z"));
+        assert_eq!(parsed.agents[0].name, "x");
     }
 }
