+++ src/config/agents/validation.rs
@@ -2,7 +2,7 @@
 //!
 //! Provides validation logic for AgentsConfig.
 
-use super::types::{AgentConfig, AgentsConfig};
+use super::AgentsConfig;
 use crate::config::ConfigError;
 
 /// Validation result type
@@ -18,13 +18,13 @@ pub fn validate_agents_config(config: &AgentsConfig) -> ValidationResult {
                 message: "Agent name cannot be empty".to_string(),
             });
         }
-        if agent.model.is_empty() {
+        if agent.model.as_ref().map_or(true, |m| m.is_empty()) {
             return Err(ConfigError::ValueError {
                 field: "model".to_string(),
                 message: format!("Agent '{}' has empty model", agent.name),
             });
         }
-        if !names.insert(&agent.name) {
+        if !names.insert(agent.name.clone()) {
             return Err(ConfigError::ValueError {
                 field: "name".to_string(),
                 message: format!("Duplicate agent name '{}'", agent.name),
@@ -34,13 +34,13 @@ pub fn validate_agents_config(config: &AgentsConfig) -> ValidationResult {
 
     // Validate parent references
     for agent in &config.agents {
-        if let Some(ref parent) = agent.parent {
-            if !names.contains(parent) {
+        if let Some(ref parent_id) = agent.parent_id {
+            if !names.contains(parent_id) {
                 return Err(ConfigError::ValueError {
                     field: "parent".to_string(),
                     message: format!(
                         "Agent '{}' references non-existent parent '{}'",
-                        agent.name, parent
+                        agent.name, parent_id
                     ),
                 });
             }
@@ -53,19 +53,22 @@ pub fn validate_agents_config(config: &AgentsConfig) -> ValidationResult {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::agent::config::AgentConfig;
+
+    fn make_agent(name: &str, model: &str) -> AgentConfig {
+        AgentConfig {
+            id: name.to_string(),
+            name: name.to_string(),
+            model: Some(model.to_string()),
+            ..Default::default()
+        }
+    }
 
     #[test]
     fn test_valid_config() {
         let config = AgentsConfig {
             version: "1.0.0".to_string(),
-            agents: vec![crate::config::agents::types::AgentConfig {
-                name: "orchestrator".to_string(),
-                model: "gpt-4".to_string(),
-                persona: "Master orchestrator".to_string(),
-                max_iterations: 100,
-                timeout_minutes: Some(60),
-                parent: None,
-            }],
+            agents: vec![make_agent("orchestrator", "gpt-4")],
         };
         validate_agents_config(&config).unwrap();
     }
@@ -74,13 +77,11 @@ mod tests {
     fn test_empty_name() {
         let config = AgentsConfig {
             version: "1.0.0".to_string(),
-            agents: vec![crate::config::agents::types::AgentConfig {
-                name: "".to_string(),
-                model: "gpt-4".to_string(),
-                persona: "".to_string(),
-                max_iterations: 100,
-                timeout_minutes: None,
-                parent: None,
+            agents: vec![AgentConfig {
+                id: String::new(),
+                name: String::new(),
+                model: Some("gpt-4".to_string()),
+                ..Default::default()
             }],
         };
         let err = validate_agents_config(&config).unwrap_err();
@@ -91,13 +92,11 @@ mod tests {
     fn test_empty_model() {
         let config = AgentsConfig {
             version: "1.0.0".to_string(),
-            agents: vec![crate::config::agents::types::AgentConfig {
+            agents: vec![AgentConfig {
+                id: "agent1".to_string(),
                 name: "agent1".to_string(),
-                model: "".to_string(),
-                persona: "".to_string(),
-                max_iterations: 100,
-                timeout_minutes: None,
-                parent: None,
+                model: None,
+                ..Default::default()
             }],
         };
         let err = validate_agents_config(&config).unwrap_err();
@@ -109,22 +108,8 @@ mod tests {
         let config = AgentsConfig {
             version: "1.0.0".to_string(),
             agents: vec![
-                crate::config::agents::types::AgentConfig {
-                    name: "agent1".to_string(),
-                    model: "gpt-4".to_string(),
-                    persona: "".to_string(),
-                    max_iterations: 100,
-                    timeout_minutes: None,
-                    parent: None,
-                },
-                crate::config::agents::types::AgentConfig {
-                    name: "agent1".to_string(),
-                    model: "claude-3-opus".to_string(),
-                    persona: "".to_string(),
-                    max_iterations: 100,
-                    timeout_minutes: None,
-                    parent: None,
-                },
+                make_agent("agent1", "gpt-4"),
+                make_agent("agent1", "claude-3-opus"),
             ],
         };
         let err = validate_agents_config(&config).unwrap_err();
@@ -136,12 +121,11 @@ mod tests {
         let config = AgentsConfig {
             version: "1.0.0".to_string(),
             agents: vec![AgentConfig {
+                id: "coder".to_string(),
                 name: "coder".to_string(),
-                model: "claude-3-opus".to_string(),
-                persona: "Coder".to_string(),
-                max_iterations: 100,
-                timeout_minutes: Some(60),
-                parent: Some("nonexistent".to_string()),
+                model: Some("claude-3-opus".to_string()),
+                parent_id: Some("nonexistent".to_string()),
+                ..Default::default()
             }],
         };
         let err = validate_agents_config(&config).unwrap_err();
