+++ src/session/bootstrap/loader.rs
@@ -29,8 +29,11 @@ use std::collections::HashMap;
 use std::io;
 use std::path::Path;
 
+use serde::{Deserialize, Serialize};
+
 /// Bootstrap 文件集合模式
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
 pub enum BootstrapMode {
     /// 运行时必须的身份/工具文件，token 消耗最小
     Minimal,
